+++ src/main/kotlin/de/shyim/shopware6/action/generator/ActionUtil.kt
@@ -1,12 +1,14 @@
 package de.shyim.shopware6.action.generator
 
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.actionSystem.LangDataKeys
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.fileTypes.LanguageFileType
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.Messages;
 import com.intellij.openapi.util.io.StreamUtil;
+import com.intellij.psi.PsiDirectory
 import com.intellij.psi.PsiFileFactory
 import com.intellij.psi.codeStyle.CodeStyleManager
 import java.io.IOException
@@ -14,14 +16,20 @@ import java.io.IOException
 
 class ActionUtil {
     companion object {
+        fun getViewDirectory(dataContext: DataContext): PsiDirectory? {
+            val view = LangDataKeys.IDE_VIEW.getData(dataContext) ?: return null
+            val directories = view.directories
+            if (directories.isEmpty()) {
+                return null
+            }
+            return directories[0] ?: return null
+        }
+
         fun buildFile(event: AnActionEvent, project: Project, templatePath: String, fileName: String?, fileType: LanguageFileType) {
             val dataContext = event.dataContext
             val view = LangDataKeys.IDE_VIEW.getData(dataContext) ?: return
-            val directories = view.directories
-            if (directories.isEmpty()) {
-                return
-            }
-            val initialBaseDir = directories[0] ?: return
+            val initialBaseDir = this.getViewDirectory(dataContext) ?: return
+
             if (initialBaseDir.findFile(fileName!!) != null) {
                 Messages.showInfoMessage("File exists", "Error")
                 return