+++ src/main/resources/fileTemplates/vue/component/index.js
@@ -1,0 +1,8 @@
+import template from './#COMPONENT-NAME#.html.twig';
+import './#COMPONENT-NAME#.scss';
+
+const { Component } = Shopware;
+
+Component.register('#COMPONENT-NAME#', {
+    template,
+});